pyproject.toml
[project]
name = "career-conversation-ai-agent-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.29.1",
    "openai>=1.79.0",
    "openai-agents>=0.0.15",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
